[Issue] Parse wrong value from float value
M: Using double parse instead of float parse

+++ Pharmacy/Pharmacy/Implement/Windows/MainScreenWindow/Action/Types/Pages/WarehouseManagementPage/AddWarehouseImportPage/MSW_WHMP_AWIP_AddMedicineToListButtonAction.cs
@@ -2,6 +2,7 @@
 using Pharmacy.Base.Utils;
 using Pharmacy.Implement.Windows.MainScreenWindow.MVVM.ViewModels.Pages.WarehouseManagementPage.OVs;
 using System;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Windows.Controls;
@@ -31,7 +32,7 @@ namespace Pharmacy.Implement.Windows.MainScreenWindow.Action.Types.Pages.Warehou
                 {
                     error.Append("Giá thuốc không hợp lệ!").AppendLine();
                 }
-                if (float.Parse(AWIPViewModel.MedicineQuantity) <= 0)
+                if (double.Parse(AWIPViewModel.MedicineQuantity) <= 0)
                 {
                     error.Append("Số lượng thuốc không hợp lệ!").AppendLine();
                 }
@@ -48,7 +49,7 @@ namespace Pharmacy.Implement.Windows.MainScreenWindow.Action.Types.Pages.Warehou
                     item.MedicineID = AWIPViewModel.SelectedMedicine.MedicineID;
                     item.MedicineName = AWIPViewModel.SelectedMedicine.MedicineName;
                     item.MedicineUnitName = AWIPViewModel.SelectedMedicine.tblMedicineUnit.MedicineUnitName;
-                    item.Quantity = float.Parse(AWIPViewModel.MedicineQuantity);
+                    item.Quantity = double.Parse(AWIPViewModel.MedicineQuantity);
                     item.UnitPrice = AWIPViewModel.MedicinePrice;
                     item.TotalPrice = (decimal)item.Quantity * item.UnitPrice;
 
@@ -57,7 +58,7 @@ namespace Pharmacy.Implement.Windows.MainScreenWindow.Action.Types.Pages.Warehou
                 else
                 {
                     item.UnitPrice = AWIPViewModel.MedicinePrice;
-                    item.Quantity += float.Parse(AWIPViewModel.MedicineQuantity);
+                    item.Quantity += double.Parse(AWIPViewModel.MedicineQuantity);
                     item.TotalPrice = (decimal)item.Quantity * item.UnitPrice;
                     dataGrid.Items.Refresh();
                 }
